pull out KavaNodeRunner into kvtool file
